(Refactor) move viewport height logic to App component
