+++ src/main.tsx
@@ -4,24 +4,14 @@ import { QueryClientProvider } from '@tanstack/react-query';
 import { queryClient } from './lib/react-query-client';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 
-import { RouterProvider } from 'react-router-dom';
-import { router } from './router';
+import App from './App';
 
 import './index.css';
-
-function setViewportHeightVar() {
-  const vh = window.innerHeight * 0.01;
-  document.documentElement.style.setProperty('--vh', `${vh}px`);
-}
-
-window.addEventListener('resize', setViewportHeightVar);
-window.addEventListener('orientationchange', setViewportHeightVar);
-setViewportHeightVar();
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <QueryClientProvider client={queryClient}>
-      <RouterProvider router={router} />
+      <App />
       {import.meta.env.DEV && <ReactQueryDevtools initialIsOpen={false} />}
     </QueryClientProvider>
   </StrictMode>,
